Switched physical resolution metadata from single to double precision floats.

+++ Extensions/LibTiff/ImagingTiffLib.pas
@@ -434,7 +434,7 @@ var
 
   procedure SaveMetadata(Tiff: PTiff; TiffPage: Integer);
   var
-    XRes, YRes: Single;
+    XRes, YRes: Double;
     ResUnit: TResolutionUnit;
     TiffResUnit, StoredTiffResUnit: Word;
   begin
